gitbasecode/requirements.txt
strictyaml>=1.3.2,<2.0.0
ruamel.yaml>=0.16.12,<1.0.0
uvicorn
fastapi
numpy
pandas
pydantic
scikit-learn
requests
joblib
transformers==4.41.2
datasets 
evaluate==0.4.2
pyarrow==16.1.0
accelerate
Pillow
peft
dvc
dvc-s3
openpyxl